Refactoring: Define utility test function globally
