+++ glean-core/python/tests/conftest.py
@@ -3,7 +3,9 @@
 # file, You can obtain one at http://mozilla.org/MPL/2.0/.
 
 
+import json
 import logging
+import os
 import socket
 import time
 from pathlib import Path
@@ -108,6 +110,37 @@ def wait_for_server(httpserver, timeout=30):
         except socket.error:
             if time.time() - start_time > timeout:
                 raise TimeoutError()
+
+
+class Helpers:
+    @staticmethod
+    def wait_for_ping(path, timeout=2) -> (str, str):
+        """
+        Wait for a ping to appear in `path` for at most `timeout` seconds.
+
+        Raises a `TimeoutError` if the file doesn't exist within the timeout.
+
+        Returns a tuple of (url path, payload).
+        """
+
+        start_time = time.time()
+        while not path.exists():
+            time.sleep(0.1)
+            if time.time() - start_time > timeout:
+                raise TimeoutError(f"No ping appeared in {path} within {timeout} seconds")
+
+        with path.open("r") as fd:
+            url_path = fd.readline()
+            serialized_ping = fd.readline()
+            payload = json.loads(serialized_ping)
+
+        os.remove(path)
+        return (url_path, payload)
+
+
+@pytest.fixture
+def helpers():
+    return Helpers
 
 
 # Setup a default webserver that pings will go to by default, so we don't hit
